requirements.txt
google-api-python-client==2.88.0
google-auth-oauthlib==1.0.0
streamlit==1.22.0
pandas==2.0.1
plotly==5.14.1
python-dotenv==1.0.0
cassandra-driver==3.25.0
pyyaml==6.0 